# DoTiles render performance optimisation